Override the rails restart command
Previously we always overrode this using the Stack Settings custom json.
As of opsworks chef 11.10, this attribute is declared as "normal"
instead of as "default", so the stack settings will not be used.

This changes our stacks to always use the restart command. It's
something we've used on every stack we've made, and is a generally
better default.