Fix permissions of created logrotate setting files
Permissions for these files should be `0644` but it is set to `0440` so
we need to override it.
Interestingly they also changed default to `0644` in later versions.
We have to use this old version because later versions don't support
the chef version we are using in opsworks which also happens to be pretty
old one.